Moved reasoning utils to protocol.
